New commit
Everything is working well right now.
The only issue is that when goal directness = 100% and Tadros start at
set points by the edge of the pool, it’s always gonna turn out the same
way.
Interestingly, Sg as it’s calculated right now increases with iteration
only when goal directness is high, not when it’s low.